[CONFIG] [REFACTOR] moved all left unsorted config parts out

+++ src/CryptoNoteCore/DataBaseConfig.h
@@ -15,8 +15,8 @@ namespace CryptoNote {
 class DataBaseConfig {
  public:
   DataBaseConfig();
-  bool init(const std::string dataDirectory, const int backgroundThreads, const int maxOpenFiles,
-            const int writeBufferSize, const int readCacheSize);
+  bool init(const std::string dataDirectory, const uint16_t backgroundThreads, const uint16_t maxOpenFiles,
+            const uint64_t writeBufferSize, const uint64_t readCacheSize);
 
   bool isConfigFolderDefaulted() const;
   std::string getDataDir() const;
